Fix XAML binding issues and control references
- Move PresetInfo class to Models directory
- Add proper namespace declarations in XAML files
- Fix async method warnings
- Resolve InitializeComponent and control reference errors
- Plugin Installation Wizard and Preset Manager now build successfully
